Feature

+++ server/src/main/java/cn/edu/sdu/orz/bug/service/FeatureService.java
@@ -3,12 +3,12 @@ package cn.edu.sdu.orz.bug.service;
 import cn.edu.sdu.orz.bug.dto.FeatureDTO;
 import cn.edu.sdu.orz.bug.entity.Feature;
 import cn.edu.sdu.orz.bug.repository.FeatureRepository;
-import cn.edu.sdu.orz.bug.vo.FeatureQueryVO;
+import cn.edu.sdu.orz.bug.utils.Utils;
 import cn.edu.sdu.orz.bug.vo.FeatureUpdateVO;
 import cn.edu.sdu.orz.bug.vo.FeatureVO;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
 
 import java.util.NoSuchElementException;
@@ -17,32 +17,53 @@ import java.util.NoSuchElementException;
 public class FeatureService {
 
     @Autowired
+    private UserService userService;
+
+    @Autowired
+    private ModuleService moduleService;
+
+    @Autowired
     private FeatureRepository featureRepository;
 
-    public String save(FeatureVO vO) {
-        Feature bean = new Feature();
-        BeanUtils.copyProperties(vO, bean);
-        bean = featureRepository.save(bean);
-        return bean.getId();
+    public boolean create(FeatureVO vO, HttpSession session) {
+        if (userService.isNotLoggedIn(session))
+            return false;
+        try {
+            Feature bean = new Feature();
+            BeanUtils.copyProperties(vO, bean, Utils.getNullPropertyNames(vO));
+            bean.setId(newID());
+            bean.setModule(moduleService.requireOne(vO.getModule()));
+            featureRepository.save(bean);
+        } catch (Exception e) {
+            return false;
+        }
+        return true;
     }
 
-    public void delete(String id) {
-        featureRepository.deleteById(id);
+    public boolean modify(FeatureUpdateVO vO, HttpSession session) {
+        if (userService.isNotLoggedIn(session))
+            return false;
+        try {
+            Feature bean = requireOne(vO.getId());
+            BeanUtils.copyProperties(vO, bean, Utils.getNullPropertyNames(vO));
+            if (vO.getOwner() != null)
+                bean.setOwner(userService.requireOne(vO.getOwner()));
+            featureRepository.save(bean);
+        } catch (Exception e) {
+            return false;
+        }
+        return true;
     }
 
-    public void update(String id, FeatureUpdateVO vO) {
-        Feature bean = requireOne(id);
-        BeanUtils.copyProperties(vO, bean);
-        featureRepository.save(bean);
-    }
-
-    public FeatureDTO getById(String id) {
-        Feature original = requireOne(id);
-        return toDTO(original);
-    }
-
-    public Page<FeatureDTO> query(FeatureQueryVO vO) {
-        throw new UnsupportedOperationException();
+    public boolean remove(String id, HttpSession session) {
+        if (userService.isNotLoggedIn(session))
+            return false;
+        try {
+            featureRepository.deleteById(id);
+        } catch (Exception e) {
+            return false;
+        }
+        return true;
     }
 
     private FeatureDTO toDTO(Feature original) {
@@ -54,5 +75,14 @@ public class FeatureService {
     public Feature requireOne(String id) {
         return featureRepository.findById(id)
                 .orElseThrow(() -> new NoSuchElementException("Resource not found: " + id));
+    }
+
+    private String newID() {
+        while (true) {
+            String id = Utils.newRandomID();
+            if (!featureRepository.existsById(id)) {
+                return id;
+            }
+        }
     }
 }
